+++ workspace/react-addon/src/agents/superagent.js
@@ -1,4 +1,4 @@
 import request from 'superagent';
 import { Runtime } from '@genx/react';
 
-Runtime.register('httpAgentCreator', () => request.agent());
+Runtime.registerCreator('httpAgent', () => request.agent());