+++ workspace/react-addon/src/agents/supermock.js
@@ -4,7 +4,7 @@ import { Runtime } from '@genx/react';
 
 const logging = (log) => Runtime.log('verbose', ['mock', log]);
 
-Runtime.register('httpAgentCreator', (config) => {
+Runtime.registerCreator('httpAgent', (config) => {
     superagentMock(request, config, logging);
     return request;
 });