+++ workspace/react-addon/src/navigations/bottomTabMaterial.js
@@ -1,5 +1,5 @@
 import { Runtime } from '@genx/react';
 import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
 
-Runtime.register('navigation:bottomTabMaterial', createMaterialBottomTabNavigator);
-Runtime.register('navigation:bottomTab', createMaterialBottomTabNavigator, true);
+Runtime.register('nav:bottomTabMaterial', createMaterialBottomTabNavigator);
+Runtime.register('nav:bottomTab', createMaterialBottomTabNavigator, true);