+++ workspace/react-addon/src/navigations/bottomTabNative.js
@@ -1,5 +1,5 @@
 import { Runtime } from '@genx/react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-Runtime.register('navigation:bottomTabNative', createBottomTabNavigator);
-Runtime.register('navigation:bottomTab', createBottomTabNavigator);
+Runtime.register('nav:bottomTabNative', createBottomTabNavigator);
+Runtime.register('nav:bottomTab', createBottomTabNavigator);