+++ workspace/react-addon/src/navigations/drawerReanimated.js
@@ -1,5 +1,5 @@
 import { Runtime } from '@genx/react';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 
-Runtime.register('navigation:drawerReanimated', createDrawerNavigator);
-Runtime.register('navigation:drawer', createDrawerNavigator, true);
+Runtime.register('nav:drawerReanimated', createDrawerNavigator);
+Runtime.register('nav:drawer', createDrawerNavigator, true);