+++ workspace/react-addon/src/navigations/stackReact.js
@@ -1,5 +1,5 @@
 import { Runtime } from '@genx/react';
 import { createStackNavigator } from '@react-navigation/stack';
 
-Runtime.register('navigation:stackReact', createStackNavigator);
-Runtime.register('navigation:stack', createStackNavigator, true);
+Runtime.register('nav:stackReact', createStackNavigator);
+Runtime.register('nav:stack', createStackNavigator, true);