+++ workspace/react-addon/src/navigations/topTabMaterial.js
@@ -1,5 +1,5 @@
 import { Runtime } from '@genx/react';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 
-Runtime.register('navigation:topTabMaterial', createMaterialTopTabNavigator);
-Runtime.register('navigation:topTab', createMaterialTopTabNavigator, true);
+Runtime.register('nav:topTabMaterial', createMaterialTopTabNavigator);
+Runtime.register('nav:topTab', createMaterialTopTabNavigator, true);