+++ workspace/react-addon/src/overlays/initFlashMessage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import FlashMessage, { showMessage, hideMessage } from 'react-native-flash-message';
 import { Runtime } from '@genx/react';
 
-Runtime.register('flash-message', { showMessage, hideMessage });
+Runtime.registerObject('messenger', { showMessage, hideMessage });
 
 export default function initialize(props) {
     Runtime.composeScreen((elScreen) => (