+++ workspace/react-addon/src/web/enableNativeView.js
@@ -1,6 +1,0 @@
-import { Runtime } from '@genx/react';
-import NativeView from '../components/NativeView';
-
-// ----------------------------------------------------------------------
-
-Runtime.register('NativeView', NativeView);