+++ workspace/react/src/formatters/index.js
@@ -1,3 +1,3 @@
-export * from './formatNumberIntl';
+export * from './formatIntl';
 export * from './formatNumber';
 export * from './formatTime';